input fix for eth. parsing fix for contract addresses
