--- tests/src/eth/kelp/test_depositAsset.py
@@ -1,19 +1,0 @@
-from web3 import Web3
-from tests.utils import run_test, load_contract
-
-contract = load_contract(
-    "0x62De59c08eB5dAE4b7E6F7a8cAd3006d6965ec16"
-)
-
-def test_kelp_depositAsset(backend, firmware, navigator, test_name, wallet_addr):
-    data = "0x02f9015401820221850746f00bd9850746f00bd98307066094036676389e48133b63a802f8635ad39e752d375d80b8e4c3ae1766000000000000000000000000a35b1b31ce002fbf2058d22f30f95d405200a15b0000000000000000000000000000000000000000000000001b340783b0b8800000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000080000000000000000000000000000000000000000000000000000000000000004065313663623065316333353133303164333338656661343334656531646166336631393335386363633363666161643163386264323733663736633133313561c001a0aec13856dbebd5a5dd1cca745781942327cb6a279f6276a8ec19c3c402294cd1a035b6f92a682f56ae9187eb84c3340c15f69649fdbbe396b5eaf7d16bb2826503"
-    run_test(
-        contract, 
-        data, 
-        backend, 
-        firmware, 
-        navigator, 
-        test_name, 
-        wallet_addr
-    )
-
