--- tests/src/eth/kelp/test_depositETH.py
@@ -1,19 +1,0 @@
-from web3 import Web3
-from tests.utils import run_test, load_contract
-
-contract = load_contract(
-    "0x62De59c08eB5dAE4b7E6F7a8cAd3006d6965ec16"
-)
-
-def test_kelp_depositETH(backend, firmware, navigator, test_name, wallet_addr):
-    data = "0x02f90138018083902d038509f5d5b7778307a4db94036676389e48133b63a802f8635ad39e752d375d880d0b8d0508de0000b8c472c51c0b000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000400000000000000000000000000000000000000000000000000000000000000042307836613137346533306162376238313262663631326462383932663038303634656364666161613638636166353034633330326630353236663066303930643836000000000000000000000000000000000000000000000000000000000000c080a0c451fdeb418249686444110516b6872489fdb57e5623e19e459a9f9dcb24487ba070a9fed5d2660b874f3b80bc14abb38d92863b1332c05dc9d11e6beb78d91034"
-    run_test(
-        contract, 
-        data, 
-        backend, 
-        firmware, 
-        navigator, 
-        test_name, 
-        wallet_addr
-    )
-
