+++ tests/src/eth/kelp/test_kelp_depositAsset.py
@@ -1,0 +1,19 @@
+from web3 import Web3
+from tests.utils import run_test, load_contract
+
+contract = load_contract(
+    "0x036676389e48133B63a802f8635AD39E752D375D"
+)
+
+def test_kelp_depositAsset(backend, firmware, navigator, test_name, wallet_addr):
+    data = "0xc3ae1766000000000000000000000000a35b1b31ce002fbf2058d22f30f95d405200a15b0000000000000000000000000000000000000000000000001b340783b0b8800000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000080000000000000000000000000000000000000000000000000000000000000004065313663623065316333353133303164333338656661343334656531646166336631393335386363633363666161643163386264323733663736633133313561"
+    run_test(
+        contract, 
+        data, 
+        backend, 
+        firmware, 
+        navigator, 
+        test_name, 
+        wallet_addr
+    )
+
