+++ tests/src/eth/kelp/test_kelp_depositETH.py
@@ -1,0 +1,19 @@
+from web3 import Web3
+from tests.utils import run_test, load_contract
+
+contract = load_contract(
+    "0x036676389e48133B63a802f8635AD39E752D375D"
+)
+
+def test_kelp_depositETH(backend, firmware, navigator, test_name, wallet_addr):
+    data = "0x72c51c0b000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000400000000000000000000000000000000000000000000000000000000000000042307836613137346533306162376238313262663631326462383932663038303634656364666161613638636166353034633330326630353236663066303930643836000000000000000000000000000000000000000000000000000000000000"
+    run_test(
+        contract, 
+        data, 
+        backend, 
+        firmware, 
+        navigator, 
+        test_name, 
+        wallet_addr
+    )
+
