+++ tests/src/eth/kelp/test_kelp_initiateWithdrawalETH.py
@@ -6,7 +6,7 @@ contract = load_contract(
 )
 
 def test_kelp_initiateWithdrawalETH(backend, firmware, navigator, test_name, wallet_addr):
-    data = "0x02f8b00150839402a08501ccf6aae1830a78d19462de59c08eb5dae4b7e6f7a8cad3006d6965ec1680b844c8393ba9000000000000000000000000eeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeee000000000000000000000000000000000000000000000000015f021397cf0000c001a0b812d5b2183bd21ccde3765da1ad02a7233f835b01bc3519db9a123f84acbf38a0761578a1e66af567e40dcf4597af674956381fc85db0f2c9881f3f95eb863726"
+    data = "0xc8393ba9000000000000000000000000eeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeee000000000000000000000000000000000000000000000000015f021397cf0000"
     run_test(
         contract, 
         data,
