+++ tests/src/polygon/test_polygon_request_maticx_swap.py
@@ -2,7 +2,7 @@ from web3 import Web3
 from tests.utils import run_test, load_contract
 
 contract = load_contract(
-    "0xfd225c9e6601c9d38d8f98d8731bf59efcf8c0e3"
+    "0xfd225C9e6601C9d38d8F98d8731BF59eFcF8C0E3"
 )
 
 def test_polygon_request_maticx_swap(backend, firmware, navigator, test_name, wallet_addr):
